pyproject.toml
[project]
name = "ava-whatsapp-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "duckdb>=1.5.1",
    "elevenlabs>=2.40.0",
    "groq>=0.37.1",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "langgraph-checkpoint-duckdb>=2.0.2",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "numpy>=2.4.3",
    "pydantic-settings>=2.13.1",
    "qdrant-client>=1.17.1",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.55.0",
    "supabase>=2.28.3",
    "torch>=2.11.0",
]

[project.scripts]
run-ui = "ai_companion.cli:run_ui"
run-server = "ai_companion.cli:run_server"
